requirements.txt
streamlit==1.43.2
tensorflow==2.12.0
opencv-python-headless
numpy
Pillow
streamlit-webrtc
